requirements.txt
python-dotenv==1.0.0
PyYAML==6.0
email-validator==1.3.1
python-box==7.0.1
pydantic==1.10.9
